renamed Message -> ChatMessage
